Project01_DTL

+++ admin/pages/modules/sanpham/viewproduct.php
@@ -19,8 +19,9 @@
         <table class="table table-bordered">
             <thead>
                 <tr>
-                    <th><a href="?quanly=sanpham&method=createProduct">Thêm sản phẩm</a></th>
-                    <th colspan="3">
+                    <th></th>
+                    <th collspan='2'><a href="?quanly=sanpham&method=createProduct">Thêm sản phẩm</a></th>
+                    <th colspan="4">
                         <input type="text" name="textBox" placeholder="Nhập vào dữ liệu muốn tìm...">
                         <input type="submit" name="searchBtn" class="searchBtn" value="Tìm kiếm"><br>
                         <?php
@@ -33,15 +34,19 @@
                 <tr>
                     <th>Stt</th>
                     <th>Mã sản phẩm</th>
+                    <th>Ảnh sản phẩm</th>
+                    <th>Màu sản phẩm</th>
                     <th>Tên sản phẩm</th>
                     <th>Mã danh mục</th>
                     <th>Số lượng</th>
                     <th>Giá sản phẩm</th>
                     <th>Trạng thái</th>
+                    <th>Nội dung sản phẩm</th>
                     <th>Chức năng</th>
                 </tr>
             </thead>
             <tbody>
+                
                 <?php
                 if (($result->num_rows) > 0) {
                     $stt=0;
@@ -49,20 +54,23 @@
                         $status = $row["tinhtrang_sanpham"] == 1 ? "Kinh doanh" : "Ngừng kinh doanh";
                         $stt++;
                         echo "
-                                    <tr>
-                                        <td>{$stt}</td>
-                                        <td>{$row["id_sanpham"]}</td>
-                                        <td>{$row["ten_sanpham"]}</td>
-                                        <td>{$row["id_danhmuc"]}</td>
-                                        <td>{$row["soluong"]}</td>
-                                        <td>{$row["gia_sanpham"]}</td>
-                                        <td>{$status}</td>
-                                        <td>
-                                            <a href='?quanly=sanpham&deleteProduct={$row["id_sanpham"]}'>Xóa</a> || 
-                                            <a href='?quanly=sanpham&method=edit&editProduct={$row["id_sanpham"]}'>Sửa</a>
-                                        </td>
-                                    </tr>
-                                    ";
+                                <tr>
+                                    <td>{$stt}</td>
+                                    <td>{$row["id_sanpham"]}</td>
+                                    <td><img src='uploads/{$row["anh_sanpham"]}' width='150px'></td>
+                                    <td><img src='uploads/{$row["mau_sanpham"]}' width='60px'</td>
+                                    <td>{$row["ten_sanpham"]}</td>
+                                    <td>{$row["id_danhmuc"]}</td>
+                                    <td>{$row["soluong"]}</td>
+                                    <td>{$row["gia_sanpham"]}</td>
+                                    <td>{$status}</td>
+                                    <td>{$row["noidung_sanpham"]}</td>
+                                    <td>";?>
+                                        <a href='?quanly=sanpham&deleteProduct=<?php echo $row["id_sanpham"];?>' onclick="if(confirm('Bạn có muốn xóa không?')){return true;}else{return false;}">Xóa</a> || 
+                                        <a href='?quanly=sanpham&method=edit&editProduct=<?php echo $row["id_sanpham"];?>'>Sửa</a>
+                                    </td>
+                                </tr>
+                <?php
                     }
                 } else {
                     echo "<h2 align='center' style='color: Tomato;'>Không có dữ liệu nào được nhập vào!</h2><br>";
